♻️ Refactor(shared) : [코드리뷰 반영] App.js return 컴포넌트 구조 수정
- 기존에 style을 위해 컴포넌트 겉에 있던 태그들을 컴포넌트 안으로 넣었습니다.
- SearchBar와 CardList를 SharedContent라는 새로운 컴포넌트를 만들어 그 안에 넣었습니다.
