+++ src/components/Footer.js
@@ -12,27 +12,29 @@ const INSTAGRAM_URL = 'https://www.instagram.com/';
 
 function Footer() {
   return (
-    <div className="footer_container">
-      <div className="sources">©codeit - 2023</div>
-      <div className="ppf">
-        <a href="/privacy">Privacy Policy</a>
-        <a href="/faq">FAQ</a>
+    <footer>
+      <div className="footer_container">
+        <div className="sources">©codeit - 2023</div>
+        <div className="ppf">
+          <a href="/privacy">Privacy Policy</a>
+          <a href="/faq">FAQ</a>
+        </div>
+        <div className="link_icon">
+          <a href={FACEBOOK_URL} target="_blank" rel="noopener noreferrer">
+            <img alt="facebook 바로가기" src={facebookIcon} />
+          </a>
+          <a href={TWITTER_URL} target="_blank" rel="noopener noreferrer">
+            <img alt="twitter 바로가기" src={twitterIcon} />
+          </a>
+          <a href={YOUTUBE_URL} target="_blank" rel="noopener noreferrer">
+            <img alt="youtube 바로가기" src={youtubeIcon} />
+          </a>
+          <a href={INSTAGRAM_URL} target="_blank" rel="noopener noreferrer">
+            <img alt="instagram 바로가기" src={instagramIcon} />
+          </a>
+        </div>
       </div>
-      <div className="link_icon">
-        <a href={FACEBOOK_URL} target="_blank" rel="noopener noreferrer">
-          <img alt="facebook 바로가기" src={facebookIcon} />
-        </a>
-        <a href={TWITTER_URL} target="_blank" rel="noopener noreferrer">
-          <img alt="twitter 바로가기" src={twitterIcon} />
-        </a>
-        <a href={YOUTUBE_URL} target="_blank" rel="noopener noreferrer">
-          <img alt="youtube 바로가기" src={youtubeIcon} />
-        </a>
-        <a href={INSTAGRAM_URL} target="_blank" rel="noopener noreferrer">
-          <img alt="instagram 바로가기" src={instagramIcon} />
-        </a>
-      </div>
-    </div>
+    </footer>
   );
 }
