+++ src/styles/folderInfo.css
@@ -1,3 +1,9 @@
+.folder_info {
+  background-color: var(--bg);
+  width: 100%;
+  padding: 80px 0px 60px;
+}
+
 .folder_info_container {
   margin: 0 auto;
   width: 210px;
@@ -33,6 +39,9 @@
 }
 
 @media (max-width: 767px) {
+  .folder_info {
+    padding: 80px 0px 40px;
+  }
   .folder_info_container {
     gap: 10px;
   }
