+++ src/styles/sharedContent.css
@@ -1,0 +1,22 @@
+main {
+  margin: 40px 0px;
+}
+
+.content_container {
+  width: 1060px;
+  margin: 0 auto;
+}
+
+@media (max-width: 1124px) {
+  .content_container {
+    width: 704px;
+  }
+}
+
+/* Mobile */
+@media (max-width: 767px) {
+  .content_container {
+    width: 100%;
+    padding: 0 32px;
+  }
+}
